requirements.txt
anyio==4.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
docker==7.0.0
h11==0.14.0
httpcore==1.0.4
httpx==0.25.2
idna==3.6
ollama==0.1.7
packaging==23.2
requests==2.31.0
sniffio==1.3.1
urllib3==2.2.1

lancedb~=0.6.2
pygame~=2.5.2
PyAudio~=0.2.14
PyYAML~=6.0.1
PyQt5~=5.15.10
tqdm~=4.66.2